fix: Refactor the code.
