+++ src/clients/index.ts
@@ -1,3 +1,4 @@
 export * from './discord';
 export * from './express';
 export * from './astar';
+export * from './requestFilter';
